requirements.txt
hopsworks
pandas
numpy
scikit-learn
requests
streamlit
python-dotenv
joblib